renamed onKeyInterpreterResult for both navigation and keyboard InterpreterResultListener to onNavKeyInterpreterResult and onKeyboardKeyInterpreterResult respectively for the sake of RecipientInputActivity where it uses both listeners and the existing method overwrote each other.
Transitioning focus between the filter text and the list view is
working by using a hack (writing a CursorWrapper that overwrites
getCount with getCount + 1 to accommodate a dummy row in the beginning
of the list to account for when the user goes back up to the filter
text from the list view, the selected item is the first Contact.  When
the user presses down next time, the selected item will be the SECOND
contact because the first Contact was still previously selected.  By
having a dummy first row, pressing down from the filter text
guarantees it will start at the first row.  

TODO: still need to figure out how to hide the first row from sight.